+++ src/gui/mod.rs
@@ -748,11 +748,16 @@ impl Application for Ajour {
                     .style(style::CatalogQueryInput(color_palette));
 
                     let catalog_query: Element<Interaction> = catalog_query.into();
+                    let catalog_source = self
+                        .config
+                        .catalog_source
+                        .map(CatalogSource::Choice)
+                        .unwrap_or(CatalogSource::None);
 
                     let source_picklist = PickList::new(
                         &mut self.catalog_search_state.sources_state,
                         &self.catalog_search_state.sources,
-                        Some(self.catalog_search_state.source),
+                        Some(catalog_source),
                         Interaction::CatalogSourceSelected,
                     )
                     .text_size(14)
@@ -885,10 +890,22 @@ impl Application for Ajour {
 
                     content = content
                         .push(catalog_query_container)
-                        .push(Space::new(Length::Fill, Length::Units(5)))
-                        .push(catalog_row_titles)
-                        .push(catalog_scrollable)
-                        .push(bottom_space)
+                        .push(Space::new(Length::Fill, Length::Units(5)));
+
+                    if self.config.catalog_source.is_none() {
+                        let status = element::status::data_container(
+                            color_palette,
+                            &localized_string("select-catalog-source-title")[..],
+                            &localized_string("select-catalog-source-description")[..],
+                            None,
+                        );
+
+                        content = content.push(status);
+                    } else {
+                        content = content.push(catalog_row_titles).push(catalog_scrollable);
+                    }
+
+                    content = content.push(bottom_space)
                 }
             }
             Mode::Settings => {
@@ -1691,7 +1708,6 @@ pub struct CatalogSearchState {
     pub category: CatalogCategory,
     pub categories: Vec<CatalogCategory>,
     pub categories_state: pick_list::State<CatalogCategory>,
-    pub source: CatalogSource,
     pub sources: Vec<CatalogSource>,
     pub sources_state: pick_list::State<CatalogSource>,
 }
@@ -1709,7 +1725,6 @@ impl Default for CatalogSearchState {
             category: Default::default(),
             categories: Default::default(),
             categories_state: Default::default(),
-            source: CatalogSource::Choice(catalog::Source::Curse),
             sources: CatalogSource::all(),
             sources_state: Default::default(),
         }
@@ -1824,6 +1839,7 @@ impl std::fmt::Display for CatalogResultSize {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub enum CatalogSource {
     Choice(catalog::Source),
+    None,
 }
 
 impl CatalogSource {
@@ -1839,6 +1855,7 @@ impl CatalogSource {
 
 impl std::fmt::Display for CatalogSource {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let empty_display_string = &localized_string("select-catalog-source-picklist")[..];
         let s = match self {
             CatalogSource::Choice(source) => match source {
                 catalog::Source::Curse => "Curse",
@@ -1847,6 +1864,7 @@ impl std::fmt::Display for CatalogSource {
                 catalog::Source::TownlongYak => "TownlongYak",
                 catalog::Source::Other => panic!("Unsupported catalog source"),
             },
+            CatalogSource::None => empty_display_string,
         };
         write!(f, "{}", s)
     }